# apps/server/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  catalog.StorageTypeCode:
    enum:
    - FRIDGE
    - FREEZER
    - KIMCHI
    - ROOM
    type: string
    x-enum-varnames:
    - StorageTypeFridge
    - StorageTypeFreezer
    - StorageTypeKimchi
    - StorageTypeRoom
...
  handlers.CreateStorageSpaceRequest:
    properties:
      name:
        example: 우리집 메인 냉장고
        type: string
      storageTypeCode:
        allOf:
        - $ref: '#/definitions/catalog.StorageTypeCode'
        description: '보관 공간 타입 코드: FRIDGE(냉장), FREEZER(냉동), KIMCHI(김치냉장), ROOM(실온)'
        enum:
        - FRIDGE
        - FREEZER
        - KIMCHI
        - ROOM
        example: FRIDGE
    type: object
  handlers.ErrorResponse:
    properties:
      message:
        example: not implemented
        type: string
    type: object
...
  handlers.UpdateMeRequest:
    properties:
      displayName:
        example: 냉리정돈고수
        type: string
    type: object
...
  handlers.UpdateStorageSpaceRequest:
    properties:
      isArchived:
        default: false
        type: boolean
      name:
        example: 작은 냉동실
        type: string
    type: object
info:
  contact: {}
  description: API for naengrijeongdon inventory management.
  title: NaengriJeongdon API
  version: 0.1.0
paths:
...
  /storage-spaces:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "501":
          description: Not Implemented
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - AccessTokenAuth: []
      summary: 보관 공간 목록 조회
      tags:
      - StorageSpaces
    post:
      consumes:
      - application/json
      description: 'storageTypeCode 코드: FRIDGE(냉장), FREEZER(냉동), KIMCHI(김치냉장), ROOM(실온).
        실시간 목록은 GET /api/v1/metadata/storage-types 참고.'
      parameters:
      - description: 생성 요청
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateStorageSpaceRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "501":
          description: Not Implemented
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - AccessTokenAuth: []
      summary: 보관 공간 생성
      tags:
      - StorageSpaces
  /storage-spaces/{spaceId}:
    delete:
      parameters:
      - description: 공간 ID
        in: path
        name: spaceId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "501":
          description: Not Implemented
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - AccessTokenAuth: []
      summary: 보관 공간 삭제
      tags:
      - StorageSpaces
    get:
      parameters:
      - description: 공간 ID
        in: path
        name: spaceId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "501":
          description: Not Implemented
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - AccessTokenAuth: []
      summary: 보관 공간 단건 조회
      tags:
      - StorageSpaces
    patch:
      consumes:
      - application/json
      parameters:
      - description: 공간 ID
        in: path
        name: spaceId
        required: true
        type: string
      - description: 수정 요청
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.UpdateStorageSpaceRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "501":
          description: Not Implemented
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - AccessTokenAuth: []
      summary: 보관 공간 수정
      tags:
      - StorageSpaces
  /users/me:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - AccessTokenAuth: []
      summary: 내 사용자 정보 조회
      tags:
      - Users
    patch:
      consumes:
      - application/json
      parameters:
      - description: 수정 요청
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.UpdateMeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "501":
          description: Not Implemented
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - AccessTokenAuth: []
      summary: 내 사용자 정보 수정
      tags:
      - Users
securityDefinitions:
  AccessTokenAuth:
    description: 'Bearer access token. Example: Bearer eyJ...'
    in: header
    name: Authorization
    type: apiKey
  RefreshTokenAuth:
    description: 'Bearer refresh token. Example: Bearer eyJ...'
    in: header
    name: X-Refresh-Token
    type: apiKey
swagger: "2.0"